pro/requirement.txt
Flask==2.0.1
pdfminer.six==20201018
openai==0.27.0
